Refactor the session actions and reducer
I changed the logic so that the session actions and reducer have
request, success and failure states. Before the API request is
a success or a failure, the isLoading slice of redux store is set
to true which will allow us to provide a better UI experience with
a loading state.